All changes

+++ src/models/borrowergroupmember.js
@@ -4,17 +4,17 @@ module.exports = (sequelize, DataTypes) => {
   const BorrowerGroupMember = sequelize.define(
     "BorrowerGroupMember",
     {
-      // Composite primary key (groupId, borrowerId)
-      groupId: { type: DataTypes.BIGINT, allowNull: false, primaryKey: true },
-      borrowerId: { type: DataTypes.BIGINT, allowNull: false, primaryKey: true },
+      // Composite PK matching the table
+      groupId:   { type: DataTypes.INTEGER, allowNull: false, primaryKey: true },
+      borrowerId:{ type: DataTypes.INTEGER, allowNull: false, primaryKey: true },
 
-      role: { type: DataTypes.STRING, allowNull: false, defaultValue: "member" },
+      role:     { type: DataTypes.STRING, allowNull: false, defaultValue: "member" },
       joinedAt: { type: DataTypes.DATE, allowNull: true },
-      leftAt: { type: DataTypes.DATE, allowNull: true },
+      leftAt:   { type: DataTypes.DATE, allowNull: true },
 
-      createdAt: { type: DataTypes.DATE, allowNull: false },
-      updatedAt: { type: DataTypes.DATE, allowNull: false },
-      deletedAt: { type: DataTypes.DATE, allowNull: true },
+      createdAt:{ type: DataTypes.DATE, allowNull: false },
+      updatedAt:{ type: DataTypes.DATE, allowNull: false },
+      deletedAt:{ type: DataTypes.DATE, allowNull: true },
     },
     {
       tableName: "BorrowerGroupMembers",
